fix: lottie error
